% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{CeeDose}
\alias{CeeDose}
\title{Get coefficients}
\usage{
CeeDose(df, id, day, slot, p, dose, ...)
}
\arguments{
\item{df}{dataset}

\item{id}{(char) id variable}

\item{day}{(char) day variable}

\item{slot}{(char) slot variable}

\item{p}{(numeric) probability of single treatment}

\item{dose}{(integer) dose}

\item{...}{model specification}
}
\value{
a vector of coefficients
}
\description{
Get coefficients
}
\examples{
# This is an internal function
}
