% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{coef.mHealthDose}
\alias{coef.mHealthDose}
\title{Get coefficients}
\usage{
\method{coef}{mHealthDose}(object)
}
\arguments{
\item{object}{mHealthDose object}
}
\value{
coefficients
}
\description{
Get coefficients
}
