% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{confint.mHealthDose}
\alias{confint.mHealthDose}
\title{Print confidence interval}
\usage{
\method{confint}{mHealthDose}(object, param, level = 0.95)
}
\arguments{
\item{object}{mHealthDose object}

\item{param}{parameters of interest}

\item{level}{confidence level}
}
\value{
custom print function for class mHealthDose
}
\description{
Print confidence interval
}
