% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mHealth.R
\name{create_corr_matrix}
\alias{create_corr_matrix}
\title{Create correlation matrix to simulate epsilon}
\usage{
create_corr_matrix(n, rho = 0.5)
}
\arguments{
\item{n}{number of observations (in HeartSteps, number of total randomization points)}

\item{rho}{correlation between adjacent observations}
}
\value{
a correlation matrix with 0.5 correlation between adjacent observations
}
\description{
Create correlation matrix to simulate epsilon
}
