% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{ee.cor.2}
\alias{ee.cor.2}
\title{Title}
\usage{
ee.cor.2(beta, y, a, h, s, p_a, cum_d, dose)
}
\arguments{
\item{beta}{initial beta}

\item{y}{(n x time points) outcomes}

\item{a}{(n x time points) treatments}

\item{h}{(n x p) baseline covariates}

\item{s}{(n x q) effect modifying covariates}

\item{p_a}{(n x time points) probability of treatment}

\item{cum_d}{(n x time points) cumulative dose}

\item{dose}{(scalar) dose of interest}
}
\value{
U (scalar) estimating equation sum
}
\description{
Title
}
