% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{ee.cor.3}
\alias{ee.cor.3}
\title{Estimating Equation}
\usage{
ee.cor.3(
  beta,
  df.wide,
  p_a,
  time,
  n.days,
  dose,
  y,
  trt,
  baseline = NULL,
  timevar = NULL,
  b.prime = NULL,
  t.prime = NULL
)
}
\arguments{
\item{beta}{(numeric) vector of coefficients}

\item{df.wide}{(data.frame) wide format of the dataset}

\item{p_a}{(matrix) probability of treatment regime}

\item{time}{(integer) number of decision points within a day}

\item{n.days}{(integer) number of total decision days}

\item{dose}{(integer) dose variable}

\item{y}{(char) outcome variable}

\item{trt}{(char) treatment variable}

\item{baseline}{(char vector) baseline variable}

\item{timevar}{(char vector) time-varying covariate}

\item{b.prime}{(char vector) baseline interaction}

\item{t.prime}{(char vector) time-varying interaction}
}
\value{
a vector of estimating equation
}
\description{
Estimating Equation
}
