% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruvka_functions.R
\name{gee_ind}
\alias{gee_ind}
\title{GEE independent estimating equation}
\usage{
gee_ind(param, Y, S, A, rho.hat, w, n, Time)
}
\arguments{
\item{param}{(numeric vector)}

\item{Y}{(numeric matrix)}

\item{S}{(numeric matrix)}

\item{A}{(numeric matrix)}

\item{rho.hat}{(numeric) Correlation}

\item{w}{(numeric matrix) Weights}

\item{n}{(numeric) Number of subjects}

\item{Time}{(numeric) Number of time points}
}
\value{
U (numeric)
}
\description{
GEE independent estimating equation
}
