% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{generate_regimes}
\alias{generate_regimes}
\title{Title generate regimes}
\usage{
generate_regimes(time_points, total_doses)
}
\arguments{
\item{time_points}{number of total randomization points within a day}

\item{total_doses}{number of doses given in a day}
}
\value{
a matrix of all possible dose combinations
}
\description{
Title generate regimes
}
