% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{get.sim.rslts}
\alias{get.sim.rslts}
\title{Get Simulated Results}
\usage{
get.sim.rslts(m = 100, dfs = dfs, print_progress = FALSE)
}
\arguments{
\item{m}{number of simulated datasets}

\item{dfs}{list of simulated datasets}

\item{print_progress}{print progress if true}
}
\value{
matrix with 2 rows: estimated mean, estimated standard deviation
}
\description{
Get Simulated Results
}
\examples{
df.corr <- simMhealth(m=100, n=37, time=5, days=42,
eta=1, rho=0.5,
theta1=1, theta2=0.8,
beta10=0.5, beta11=0, beta12=0.2,
p=0.5)
rslt <- get.simulated.rslts(m=100, dfs=df.corr, print_progress=TRUE)
}
