% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{get.sim.rslts2}
\alias{get.sim.rslts2}
\title{Get Simulation Results}
\usage{
get.sim.rslts2(m, dfs, id, day, slot, p, ..., print_progress = FALSE)
}
\arguments{
\item{m}{number of simulations}

\item{dfs}{list of dataframes}

\item{id}{(char) id variable}

\item{day}{(char) day variable}

\item{slot}{(char) slot variable}

\item{p}{(numeric) probability of single treatment}

\item{...}{(char vector) model specification
variable names of \code{dose}, \code{y}, \code{trt}, \code{baseline}, \code{timevar}, \code{b.prime}, \code{t.prime}}

\item{print_progress}{(logical) print progress}
}
\value{
matrix of mean and standard deviation of estimated coefficients
}
\description{
Get Simulation Results
}
