% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{get_p_a}
\alias{get_p_a}
\title{Title get regime probability}
\usage{
get_p_a(a, p)
}
\arguments{
\item{a}{matrix of treatments (n x time points)}

\item{p}{probability of single treatment}
}
\value{
a list of cumulative dose, probability of a_t, and dose weight
}
\description{
Title get regime probability
}
