% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{mHealthDose}
\alias{mHealthDose}
\title{mHealthDose fit}
\usage{
mHealthDose(
  df,
  id,
  day,
  slot,
  p,
  dose,
  ...,
  boot = 100,
  cores = 1,
  cl = NULL,
  sim = FALSE,
  print_progress = FALSE
)
}
\arguments{
\item{df}{dataset}

\item{id}{(char) id variable}

\item{day}{(char) day variable}

\item{slot}{(char) slot variable}

\item{p}{(numeric) probability of single treatment}

\item{dose}{(integer) dose}

\item{...}{model specification
variable names of \code{dose}, \code{y}, \code{trt}, \code{baseline}, \code{timevar}, \code{b.prime}, \code{t.prime}}

\item{boot}{(integer) number of bootstrap samples}

\item{cores}{(integer) number of cores to use for bootstrap}

\item{cl}{(cluster) cluster object}

\item{sim}{(logical) true if simulation setting, false if real data}

\item{print_progress}{(logical) print progress}
}
\value{
list of coefficients and standard errors
}
\description{
mHealthDose fit
}
\examples{
m=5; n=10; time=5; days=2; eta=c(3,2,1,4); rho=0.5; theta1=c(3,2,1); theta2=0; 
beta10=1; beta11=0.5; beta12=0.3; p=0.5; dose=1
dfs <- simMhealth(m, n, time, days, eta, rho, theta1, theta2, beta10, beta11, beta12, p)
mHealthDose(dfs, id="ID", day="DAY", slot="SLOT", p,
           dose=dose, y="Y", trt="A",
           baseline=c("H1", "H2"), timevar=c("S1", "S2"),
           b.prime=c("H1"), t.prime=c("S1"),
           boot = m, cores = 2, sim = TRUE, print_progress=FALSE)
}
