% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{mHealthDoses}
\alias{mHealthDoses}
\title{wrapper for multiple doses}
\usage{
mHealthDoses(df, id, day, slot, p, doses, ...)
}
\arguments{
\item{df}{dataset}

\item{id}{(char) id variable}

\item{day}{(char) day variable}

\item{slot}{(char) slot variable}

\item{p}{(numeric) probability of single treatment}

\item{doses}{}

\item{...}{}
}
\value{
list of mHealthDose objects
}
\description{
wrapper for multiple doses
}
