% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{plot.mHealthDoses.summary}
\alias{plot.mHealthDoses.summary}
\title{Plot treatment effect}
\usage{
\method{plot}{mHealthDoses.summary}(x, y, ...)
}
\arguments{
\item{x}{mHealthDoses object}

\item{y}{varlevels}

\item{...}{not used}
}
\value{
ggplot of treatment effects
}
\description{
Plot treatment effect
}
\examples{
# fill in later
}
