% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{print.mHealthDose}
\alias{print.mHealthDose}
\title{Print mHealthDose object}
\usage{
\method{print}{mHealthDose}(x, ...)
}
\arguments{
\item{x}{mHealthDose object}

\item{...}{optional}
}
\value{
invisible
}
\description{
Print mHealthDose object
}
