% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{print.summary.mHealthDose}
\alias{print.summary.mHealthDose}
\title{Print Summary}
\usage{
\method{print}{summary.mHealthDose}(x, ...)
}
\arguments{
\item{x}{summary.mHealthDose object}

\item{...}{optional}
}
\value{
invisible
}
\description{
Print Summary
}
