% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{pval}
\alias{pval}
\title{Get p-values}
\usage{
pval(object, param)
}
\arguments{
\item{object}{mHealthDose object}

\item{param}{parameters of interest}
}
\value{
p-values
}
\description{
Get p-values
}
