% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mHealth.R
\name{simMhealth}
\alias{simMhealth}
\title{Mobile Health Data Simulation}
\usage{
simMhealth(
  m,
  n,
  time,
  days,
  eta,
  rho,
  theta1,
  theta2,
  beta10,
  beta11,
  beta12,
  p,
  print_progress = FALSE
)
}
\arguments{
\item{m}{number of replications}

\item{n}{number of observations (in HeartSteps, number of participants)}

\item{time}{number of time points (in HeartSteps, 5)}

\item{days}{number of days (in HeartSteps, 42)}

\item{eta}{effect of baseline covariate}

\item{rho}{correlation between adjacent observations}

\item{theta1}{effect of time-varying covariate}

\item{theta2}{effect of lagged treatment (lag 1)}

\item{beta10}{effect of treatment at time t}

\item{beta11}{interaction effect with some baseline covariate}

\item{beta12}{interaction effect with some time-varying covariate}

\item{p}{probability of treatment}
}
\value{
A list of data frames that contains
\itemize{
\item ID: participant ID
\item OBS: observation number
\item SLOT: time slot
\item H: baseline covariate
\item S: time-varying covariate
\item A: treatment indicator
\item A.prev: treatment indicator at lag 1
\item eps: error term
\item Y: outcome
}
}
\description{
\deqn{Y_{t+1} = \eta H + \theta_1 S_t + \theta_2 A_{t-1} + (\beta_{10} + \beta_{11}H' + \beta_{12}S_t')A_t}
}
