% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruvka_functions.R
\name{simulate.boruvka}
\alias{simulate.boruvka}
\title{Title}
\usage{
\method{simulate}{boruvka}(At_1, At, St_1, St, xi, eta1, eta2, theta1, theta2, beta10, beta11)
}
\arguments{
\item{beta11}{}
}
\description{
Title
}
