% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{summary.mHealthDoses}
\alias{summary.mHealthDoses}
\title{Title}
\usage{
\method{summary}{mHealthDoses}(object, ...)
}
\arguments{
\item{object}{mHealthDoses object}

\item{...}{not used}
}
\value{
plot dose response relationship
}
\description{
Title
}
