% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruvka_functions.R, R/simulate_mHealth.R
\name{create_corr_matrix}
\alias{create_corr_matrix}
\title{Get correlation matrix}
\usage{
create_corr_matrix(n, rho = 0.5)

create_corr_matrix(n, rho = 0.5)
}
\arguments{
\item{n}{number of observations (in HeartSteps, number of total randomization points)}

\item{rho}{correlation between adjacent observations}
}
\value{
(n x n) corr_matrix

a correlation matrix with 0.5 correlation between adjacent observations
}
\description{
Get correlation matrix

Create correlation matrix to simulate epsilon
}
