% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{mHealthDose}
\alias{mHealthDose}
\title{Get Empirical Results}
\usage{
mHealthDose(
  m,
  df,
  id,
  day,
  slot,
  y,
  trt,
  dose,
  p,
  baseline = NULL,
  timevar = NULL,
  b.prime = NULL,
  t.prime = NULL,
  print_progress = FALSE
)
}
\arguments{
\item{m}{number of simulations}

\item{id}{(char) id variable}

\item{day}{(char) day variable}

\item{slot}{(char) slot variable}

\item{y}{(char) outcome variable}

\item{trt}{(char) treatment variable}

\item{dose}{(integer) dose variable}

\item{p}{(numeric) probability of single treatment}

\item{baseline}{(char vector) baseline variable}

\item{timevar}{(char vector) time-varying covariate}

\item{b.prime}{(char vector) baseline interaction}

\item{t.prime}{(char vector) time-varying interaction}

\item{print_progress}{(logical) print progress}

\item{dfs}{list of dataframes}
}
\value{
matrix of mean and standard deviation of estimated coefficients
}
\description{
Get Empirical Results
}
