% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{mHealthDose}
\alias{mHealthDose}
\title{mHealthDose fit}
\usage{
mHealthDose(
  df,
  id,
  day,
  slot,
  p,
  ...,
  se = TRUE,
  boot = 100,
  cores = 1,
  ncpus = 1,
  cl = NULL,
  print_progress = FALSE
)
}
\arguments{
\item{df}{dataset}

\item{id}{(char) id variable}

\item{day}{(char) day variable}

\item{slot}{(char) slot variable}

\item{p}{(numeric) probability of single treatment}

\item{...}{model specification
variable names of \code{dose}, \code{y}, \code{trt}, \code{baseline}, \code{timevar}, \code{b.prime}, \code{t.prime}}

\item{se}{(logical) get standard error}

\item{boot}{(integer) number of bootstrap samples}

\item{cores}{(integer) number of cores to use for bootstrap}

\item{ncpus}{(integer) number of cores to use for bootstrap}

\item{cl}{(cluster) cluster object}

\item{print_progress}{(logical) print progress}
}
\value{
list of coefficients and standard errors
}
\description{
mHealthDose fit
}
