% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{mHealthDoses}
\alias{mHealthDoses}
\title{wrapper for multiple doses}
\usage{
mHealthDoses(df, id, day, slot, p, doses, ...)
}
\arguments{
\item{df}{dataset}

\item{id}{(char) id variable}

\item{day}{(char) day variable}

\item{slot}{(char) slot variable}

\item{p}{(numeric) probability of single treatment}

\item{doses}{(integer vector) doses}

\item{...}{model specification}
}
\value{
list of mHealthDose objects
}
\description{
wrapper for multiple doses
}
\examples{
m=5; n=10; time=5; days=5; eta=c(0,0,0,0); rho=0; theta1=c(0,0,0); theta2=0;
beta10=5; beta11=c(0,0); beta12=c(0,0); p=0.5; dose=1
dfs <- simMhealth(m, n, time, days, eta, rho, theta1, theta2, beta10, beta11, beta12, p)
fits <- mHealthDoses(dfs, id="ID", day="DAY", slot="SLOT", p,
                     doses=c(1,2,3,4,5), y="Y", trt="A",
                     baseline=c("H1", "H2"), timevar=c("S1", "S2"),
                     b.prime=c("H1", "H2"), t.prime=c("S1", "S2"),
                     boot = m, cores = 1, sim = TRUE)
}
