% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{summary.mHealthDose}
\alias{summary.mHealthDose}
\title{mHealthDose summary}
\usage{
\method{summary}{mHealthDose}(object, ...)
}
\arguments{
\item{object}{mHealthDose object}

\item{...}{summary}
}
\value{
the summary object
}
\description{
mHealthDose summary
}
