% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{summary.mHealthDoses}
\alias{summary.mHealthDoses}
\title{Title}
\usage{
\method{summary}{mHealthDoses}(x, y, ...)
}
\arguments{
\item{x}{(mHealthDoses) object}

\item{y}{}

\item{...}{}
}
\value{
plot dose response relationship
}
\description{
Title
}
