% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{summary.mHealthDoses}
\alias{summary.mHealthDoses}
\title{mHealthDoses summary}
\usage{
\method{summary}{mHealthDoses}(object, ...)
}
\arguments{
\item{object}{mHealthDoses object}

\item{...}{not used}
}
\value{
plot dose response relationship
}
\description{
mHealthDoses summary
}
\examples{
m=5; n=10; time=5; days=10; eta=c(3,2,1,4); rho=0.5; theta1=c(3,2,1); theta2=0;
beta10=5; beta11=c(3,2); beta12=c(2,1); p=0.5; dose=1
dfs <- simMhealth(m, n, time, days, eta, rho, theta1, theta2, beta10, beta11, beta12, p)
fits <- mHealthDoses(dfs, id="ID", day="DAY", slot="SLOT", p,
                     doses=c(1,2,3,4,5), y="Y", trt="A",
                     baseline=c("H1", "H2"), timevar=c("S1", "S2"),
                     b.prime=c("H1", "H2"), t.prime=c("S1", "S2"),
                     boot = m, cores = 1, sim = TRUE)
summary(fits)
}
